Seperate ContainerGenerator out of SourceGenerator
